Fixed docker deployment.

+++ spring-boot/admin/src/main/java/eu/coatrack/admin/service/GatewayConfigFilesStorage.java
@@ -49,21 +49,24 @@ public class GatewayConfigFilesStorage {
     @Value("${ygg.admin.api-base-url-for-gateway}")
     private String adminApiBaseUrlForGateway;
 
+    @Value("${is-run-via-docker-compose-setup}")
+    private boolean isDockerComposeSetup;
+
     @PostConstruct
     private void emptyProxyConfigFilesFolderIfExists() throws IOException {
-        //TODO Disable this feature for docker containers. Maybe use a variable like isDockerComposeSetup = false, which
-        //  is overridden by the docker-compose.yml.
         //TODO Make it makes sense to make a docker-compose profile (application-docker-compose.yml) instead
         //  of writing everything in the docker-compose.yml.
-        Path targetFolderPath = Paths.get(System.getProperty("user.dir") + "/target");
-        if (Files.notExists(targetFolderPath))
-            Files.createDirectory(targetFolderPath);
+        if (!isDockerComposeSetup){
+            Path targetFolderPath = Paths.get(System.getProperty("user.dir") + "/target");
+            if (Files.notExists(targetFolderPath))
+                Files.createDirectory(targetFolderPath);
 
-        Path gatewayConfigFilesFolderPath = Paths.get(gatewayConfigFilesFolderLocation);
-        if (Files.exists(gatewayConfigFilesFolderPath)){
-            FileUtils.deleteDirectory(gatewayConfigFilesFolderPath.toFile());
+            Path gatewayConfigFilesFolderPath = Paths.get(gatewayConfigFilesFolderLocation);
+            if (Files.exists(gatewayConfigFilesFolderPath)){
+                FileUtils.deleteDirectory(gatewayConfigFilesFolderPath.toFile());
+            }
+            Files.createDirectory(gatewayConfigFilesFolderPath);
         }
-        Files.createDirectory(gatewayConfigFilesFolderPath);
     }
 
     public void addGatewayConfigFile(Proxy proxy) throws IOException {
